Split callbacks in two files
